LinuxRenderer: move duplicate defines to new RendererGLCommon.h
